factor utility functions out of the pattern assembler and put 'em in our utilities module; move isPatternFile into pattern_engines module since that's where the knowledge it needs is concentrated
