+++ builder/list_item_hunter.js
@@ -16,6 +16,7 @@
 		var extend = require('util')._extend,
 				pa = require('./pattern_assembler'),
 				smh = require('./style_modifier_hunter'),
+				plutils = require('./utilities'),
 				config = require('../config.json'),
         of = require('./object_factory');
 
@@ -48,7 +49,7 @@
 
 					//check for a local listitems.json file
 					var listData = JSON.parse(JSON.stringify(patternlab.listitems));
-					listData = pattern_assembler.merge_data(listData, pattern.listitems);
+					listData = plutils.mergeData(listData, pattern.listitems);
 
 					//iterate over each copied block, rendering its contents along with pattenlab.listitems[i]
 					for(var i = 0; i < repeatedBlockTemplate.length; i++){
@@ -61,8 +62,8 @@
 						var globalData = JSON.parse(JSON.stringify(patternlab.data));
 						var localData = JSON.parse(JSON.stringify(pattern.jsonFileData));
 
-						var allData = pattern_assembler.merge_data(globalData, localData);
-						allData = pattern_assembler.merge_data(allData, itemData != undefined ? itemData[i] : {}); //itemData could be undefined if the listblock contains no partial, just markup
+						var allData = plutils.mergeData(globalData, localData);
+						allData = plutils.mergeData(allData, itemData != undefined ? itemData[i] : {}); //itemData could be undefined if the listblock contains no partial, just markup
 						allData.link = extend({}, patternlab.data.link);
 
 						//check for partials within the repeated block
