--- builder/util.js
@@ -1,22 +1,0 @@
-/*
- * patternlab-node - v0.14.0 - 2015
- *
- * Brian Muenzenmeyer, Geoffrey Pursell and the web community.
- * Licensed under the MIT license.
- *
- * Many thanks to Brad Frost and Dave Olsen for inspiration, encouragement, and advice.
- *
- */
-
-(function () {
-  var path = require('path');
-
-  var util = {
-    // takes a string of the filename or path, and will tell you if it refers to
-    isPseudoPattern: function () {
-
-    }
-  };
-
-  module.exports = util;
-}());
